refactor large file into sub components
